Draft: split monolith into src/ui/draft module (controls/table/panels/hook); add loading spinner, error banner, busy-disabled buttons
